// react-ui/types/generated/models/game-template-summary.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export interface GameTemplateSummary {
    id: string;
    name: string;
    category: string;
    isSystemTemplate: boolean;
    description: string;
    minPlayers: number;
    maxPlayers: number;
    updatedAt: Date;
}